+++ pdvalidate/__init__.py
@@ -1,0 +1,1 @@
+from pdvalidate._version import __version__