+++ tests/coverage.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/bash
+
+printf "\nSetting up ...\n"
+coverage run -m pytest
+
+printf "\n\nRunning coverage test(s) ...\n"
+coverage report -m ../pdvalidate/validation.py
+
+printf "\n\n"